feat: add invite member functionality with form and trigger components
- Implemented `InviteMemberTrigger` and `InviteMemberForm` components for inviting members to a group.
- Created `useInviteMember` hook to handle form submission and API interaction.
- Added validation for email input using Zod schema.
- Integrated toast notifications for success and error handling during the invitation process.
- Updated notifications model to include new invitation notifications.
- Enhanced user experience with loading states and reset functionality in forms.

fix: improve login and registration form validation messages

- Updated error messages for password and username fields in login and registration forms to provide clearer guidance.

refactor: optimize notification fetching and rendering

- Combined fetching of all notifications and user-specific notifications into a single query for efficiency.
- Improved notification rendering logic to handle empty states gracefully.

chore: add middleware for authentication checks

- Introduced middleware to protect routes and ensure users are authenticated before accessing certain pages.

style: enhance UI components for better responsiveness

- Adjusted styles in sidebar, header, and various widget components for improved layout and responsiveness.
- Added new table component with search functionality for better data presentation.

test: ensure all new components and hooks are covered by unit tests

- Added tests for the new invite member functionality and updated existing tests to cover recent changes.

+++ src/shared/lib/services/group-service.ts
@@ -66,9 +66,16 @@ export class GroupService {
 
 	async delete<T extends Prisma.GroupDefaultArgs>(
 		id: string,
+		userId: string,
 		args?: Prisma.SelectSubset<T, Prisma.GroupDefaultArgs>
 	): Promise<Prisma.GroupGetPayload<T>> {
 		const params = args ?? {}
+
+		const { admins: oldAdmins } = await this.findById(id)
+
+		const isAdmin = oldAdmins.some(oldAdmin => oldAdmin === userId)
+
+		if (!isAdmin) throw new ApiError(BASE_ERRORS.BadRequest, 'Only admins can delete the group')
 
 		return (await prisma.group.delete({
 			where: {
@@ -93,15 +100,90 @@ export class GroupService {
 	async update<T extends Prisma.GroupDefaultArgs>(
 		id: string,
 		payload: Prisma.GroupUncheckedUpdateInput,
+		userId: string,
 		args?: Prisma.SelectSubset<T, Prisma.GroupDefaultArgs>
 	): Promise<Prisma.GroupGetPayload<T>> {
 		const params = args ?? {}
+
+		const newAdmins = (payload.admins ?? []) as string[]
+
+		const { admins: oldAdmins } = await this.findById(id)
+
+		const isAdmin = oldAdmins.some(oldAdmin => oldAdmin === userId)
+
+		if (!isAdmin) throw new ApiError(BASE_ERRORS.BadRequest, 'Only admins can update the group')
+
+		const adminIds = Array.from(new Set([...oldAdmins, ...newAdmins]))
 
 		return (await prisma.group.update({
 			where: {
 				id
 			},
-			data: { ...payload, id },
+			data: { ...payload, admins: adminIds, id },
+			...params
+		})) as Prisma.GroupGetPayload<T>
+	}
+
+	async addMember<T extends Prisma.GroupDefaultArgs>(
+		id: string,
+		userId: string,
+		args?: Prisma.SelectSubset<T, Prisma.GroupDefaultArgs>
+	): Promise<Prisma.GroupGetPayload<T>> {
+		const params = args ?? {}
+
+		const group = await this.findById(id)
+
+		return (await prisma.group.update({
+			where: {
+				id: group.id
+			},
+			data: { developers: { connect: { id: userId } } },
+			...params
+		})) as Prisma.GroupGetPayload<T>
+	}
+
+	async kick<T extends Prisma.GroupDefaultArgs>(
+		userId: string,
+		groupId: string,
+		args?: Prisma.SelectSubset<T, Prisma.GroupDefaultArgs>
+	): Promise<Prisma.GroupGetPayload<T>> {
+		const params = args ?? {}
+
+		const group = await this.findById(groupId, { include: { developers: true } })
+
+		const admins = group.admins.filter(admin => admin !== userId)
+		const developers = group.developers.filter(developer => developer.id !== userId)
+
+		return (await prisma.group.update({
+			where: {
+				id: groupId
+			},
+			data: {
+				developers: {
+					set: developers
+				},
+				admins: [...admins]
+			},
+			...params
+		})) as Prisma.GroupGetPayload<T>
+	}
+
+	async demoteToRegular<T extends Prisma.GroupDefaultArgs>(
+		userId: string,
+		groupId: string,
+		args?: Prisma.SelectSubset<T, Prisma.GroupDefaultArgs>
+	): Promise<Prisma.GroupGetPayload<T>> {
+		const params = args ?? {}
+
+		const { admins } = await this.findById(groupId, { include: { developers: true } })
+
+		return (await prisma.group.update({
+			where: {
+				id: groupId
+			},
+			data: {
+				admins: admins.filter(admin => admin !== userId)
+			},
 			...params
 		})) as Prisma.GroupGetPayload<T>
 	}
